[TOOL-114] Parse types and literals when extracting decorators
This removes parsing logic from the upper layers and should make code a bit easier to follow.

It does mean that we lose the ability to output only the part of the code that failed to parse, and instead have to output the whole endpoint code. That's probably fine, as they should still be rather small. Arguably, it will also give useful context.