Add a `spot lint` command with two example rules (#106)
The `lint` command will help ensure that an API follows a set of reasonable requirements.

We're starting by defining linting rules based on Airtasker's API design guidelines, but the goal is to make it an extendable set over time.

The command will be invoked with:

```
spot lint contract.ts
```
The plan is to later introduce a linting configuration file, e.g. `.spotlintrc` which would be automatically found based on the contract's path. Or we could make that provided via a flag. This will be defined later.

As part of implementing the `lint` command, we need a way to access the `ContractNode`, which has useful debugging information such as the source location. The `safeParse()` function is therefore updated to return both the source node and the cleansed node.