Remove usage of Result<> wrapper in ContractMismatcher
The Result<> wrapper is a useful pattern to distinguish between the "happy path" and a possible error scenario.

In the context of ContractMismatcher, most error scenarios seem to correspond to unexpected errors. These should be impossible, because at this stage the contract has been parsed successfully. Therefore, throwing an error may be a better option as it reflects a problem with Spot itself.

This PR proposes to remove usage of the Result<> wrapper in ContractMismatcher. This would make its usage simpler, as otherwise users of the library may assume that `result.isOk()` indicates the absence of violation (a mistake I made myself).

*Before:*
```ts
const validatorResult = contractValidator.findViolations(...);
if (validatorResult.isErr()) {
  // Handle Spot internal errors.
} else {
  const validatorResultUnwrapped = validatorResult.unwrap();
  if (validatorResultUnwrapped.violations.length > 0) {
    // There are violations.
  } else {
    // We're fine!
  }
}
```

*After:*
```ts
const validatorResult = contractValidator.findViolations(...);
if (validatorResult.violations.length > 0) {
  // There are violations.
} else {
  // We're fine!
}
}
```