[PSC-1988] Fix `--proxyBaseUrl` argument to `spot mock` (#2062)
## Description, Motivation and Context

There exists functionality in `spot` to be able to run a mock server
with an optional proxy. When the proxy isn't supplied, all endpoints
that are defined in the given contact will have a stubbed response
returned, with random, structurally valid data being returned. When the
optional proxy argument is supplied, any non-draft SPOT contracts will
be proxied through to the upstream instead of having mocked data
returned.

The intent of this proxy argument is to allow you to perform local
development against a stubbed new endpoint that doesn't yet have an
implementation upstream.

When attempting to use this though, I found that the proxy was never
being used, even though the code looked like it was meant to be doing
the right thing. Upon further digging, I found that the proxy argument
was never setup to be used correctly due to a bug in the code that
TypeScript did not pick up on, due to the use of a generic object of
options being passed to an object.

This PR:
1. Fixes the aforementioned bug, and now actually passes the proxy
configuration through from the CLI parser to the mock server function
call.
2. Updates the proxying logic to set the `host` header correctly when
invoking HTTPs endpoints.

## Checklist:

- [x] I've added/updated tests to cover my changes
- [x] I've created an issue associated with this PR