[PSC-1988] Enable `spot mock` to have a 404 fallback proxy (#2066)
In order to make the mock server more useful in the case where you have
multiple different upstream servers on the same hostname, this PR adds a
new optional fallback proxy to `spot mock`, which is used instead of
instantly 404'ing the received request if the request does not match any
of the given contracts. If provided, instead of 404'ing, it will send
the request through to the fallback proxy.