[NT] Correct the hostname for proxy configurations with port numbers (#2069)
## Description, Motivation and Context

This PR fixes a bug in the proxy hostname parsing, where we currently
accidentally use the wrong field from the parsed `URL` object to refer
to the hostname of the proxy.
https://developer.mozilla.org/en-US/docs/Web/API/URL — the `host`
attribute is of the form `<host>:<port>` if present, whereas `hostname`
is just `<host>`. How we use the parsed attribute expects just a
hostname without a port number.

This PR fixes this bug by using the `hostname` attribute of the parsed
URL instead of the `host` attribute. While I was here, I also added
tests for this proxy URL parsing function given there were none before.

## Checklist:

- [x] I've added/updated tests to cover my changes
- [x] I've created an issue associated with this PR