[NT] Optional proxy server for draft endpoints (#2070)
## Description, Motivation and Context

This PR introduces a new option, `--proxyMockBaseUrl`, which can be used
to proxy requests for draft contracts to a HTTP server instead of
returning mocked response. With this, we now have the functionality to
have a local implementation of a particular new draft contract while
using existing live implementations for all other contracts and
non-contract requests. For example, what I've been using this for
locally is:

```
spot mock \
    --port 5601 \
    --proxyBaseUrl https://api.dev.company.com \
    --proxyFallbackBaseUrl https://api.dev.company.com \
    --proxyMockBaseUrl http://127.0.0.1:3000 \  # Hit a locally running server for draft requests
    ~/some/path/to/spots/api.ts
```

The diff of this PR is best viewed with whitespace ignored, as there's a
large block of indentation changes:
https://github.com/airtasker/spot/pull/2070/files?diff=split&w=1

## Checklist:

- [x] I've added/updated tests to cover my changes
- [x] I've created an issue associated with this PR