[NT] Split out `docs` dependencies (#2084)
## Description, Motivation and Context

The dependencies in the top-level `package.json` file are not
particularly representative of SPOT itself. Many of them, especially the
large external libraries (including mobx and react 😱), are only needed
for building the docs.

This PR moves all of the docs-only dependencies out to their own
`package.json` and `yarn.lock` files. By having the docs dependencies
separate, we should have a lot more confidence when upgrading external
dependencies that we're not going to break the main application.